ml_kem::algebra

Function base_case_multiply

fn base_case_multiply(
    a0: Elem<BaseField>,
    a1: Elem<BaseField>,
    b0: Elem<BaseField>,
    b1: Elem<BaseField>,
    i: usize,
) -> (Elem<BaseField>, Elem<BaseField>)

Algorithm 12: BaseCaseMultiply

This is a hot loop. We promote to u64 so that we can do the absolute minimum number of modular reductions, since these are the expensive operation.