ml_kem::decapsulation_key

Struct DecapsulationKey

pub struct DecapsulationKey<P>
where
    P: KemParams,
{
    dk_pke: MaybeBox<DecryptionKey<P>>,
    ek: EncapsulationKey<P>,
    d: Option<MaybeBox<B32>>,
    z: MaybeBox<B32>,
}

A DecapsulationKey provides the ability to generate a new key pair, and decapsulate an encapsulated shared key.

Fields
dk_pke: MaybeBox<DecryptionKey<P>>

Decryption key.

ek: EncapsulationKey<P>

Associated encapsulation key.

d: Option<MaybeBox<B32>>

Seed this key was initialized from.

z: MaybeBox<B32>

Random string used during the implicit rejection process.

Implementations

impl<P> DecapsulationKey<P>
where P: KemParams,

pub fn from_seed(seed: Seed) -> Self

Create a DecapsulationKey instance from a 64-byte random seed value.

pub fn from_expanded( enc: &ExpandedDecapsulationKey<P>, ) -> Result<Self, InvalidKey>

👎Deprecated since 0.3.0:
use DecapsulationKey::from_seed instead

Initialize a DecapsulationKey from the serialized expanded key form.

Note that this form is deprecated in practice; use DecapsulationKey::from_seed. See ExpandedKeyEncoding for more information.

Errors

Returns InvalidKey in the event the expanded key failed validation

pub fn to_seed(&self) -> Option<Seed>

Serialize the Seed value: 64-bytes which can be used to reconstruct the DecapsulationKey.

Warning!
This value is key material. Please treat it with care.

Returns

Some if the DecapsulationKey was initialized using from_seed or generate.

None if the DecapsulationKey was initialized from the expanded form.

pub fn encapsulation_key(&self) -> &EncapsulationKey<P>

Get the EncapsulationKey which corresponds to this DecapsulationKey.

pub(crate) fn try_generate_from_rng<R>( rng: &mut R, ) -> Result<Self, <R as TryRng>::Error>
where R: TryCryptoRng + ?Sized,

pub(crate) fn generate_deterministic(d: B32, z: B32) -> Self