ml_kem

Module kem

Section 6. The ML-KEM Key-Encapsulation Mechanism

Structs

DecapsulationKey

A DecapsulationKey provides the ability to generate a new key pair, and decapsulate an encapsulated shared key.

EncapsulationKey

An EncapsulationKey provides the ability to encapsulate a shared key so that it can only be decapsulated by the holder of the corresponding decapsulation key.

Kem

An implementation of overall ML-KEM functionality. Generic over parameter sets, but then ties together all of the other related types and sizes.

Traits

Decapsulate

A value that can be used to decapsulate an encapsulated key. Often, this will just be a secret key. But, as with Encapsulate, it can be a bundle of secret keys, or it can include a sender’s private key for authenticated encapsulation.

Encapsulate

A value that can be encapsulated to. Often, this will just be a public key. However, it can also be a bundle of public keys, or it can include a sender’s private key for authenticated encapsulation.