ml_kem::kem

Trait Decapsulate

pub trait Decapsulate<EK, SS> {
    type Error: Debug;

    // Required method
    fn decapsulate(&self, encapsulated_key: &EK) -> Result<SS, Self::Error>;
}

A value that can be used to decapsulate an encapsulated key. Often, this will just be a secret key. But, as with Encapsulate, it can be a bundle of secret keys, or it can include a sender’s private key for authenticated encapsulation.

Required Associated Types

type Error: Debug

Decapsulation error

Required Methods

fn decapsulate(&self, encapsulated_key: &EK) -> Result<SS, Self::Error>

Decapsulates the given encapsulated key