ml_kem

Module ml_kem_1024

ML-KEM-1024 is the parameter set for security category 5, corresponding to key search on a block cipher with a 256-bit key.

Structs

MlKem1024

MlKem1024 is the parameter set for security category 5, corresponding to key search on a block cipher with a 256-bit key.

Type Aliases

Ciphertext

Encoded ML-KEM-512 ciphertexts.

DecapsulationKey

An ML-KEM-1024 DecapsulationKey which provides the ability to generate a new key pair, and decapsulate an encapsulated shared key.

EncapsulationKey

An ML-KEM-1024 EncapsulationKey provides the ability to encapsulate a shared key so that it can only be decapsulated by the holder of the corresponding decapsulation key.