ml_kem::ml_kem_1024

Type Alias EncapsulationKey

pub type EncapsulationKey = EncapsulationKey<MlKem1024>;

An ML-KEM-1024 EncapsulationKey provides the ability to encapsulate a shared key so that it can only be decapsulated by the holder of the corresponding decapsulation key.

Aliased Type

pub struct EncapsulationKey {
    ek_pke: MaybeBox<EncryptionKey<MlKem1024>>,
    h: Array<u8, UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>>,
}

Fields
ek_pke: MaybeBox<EncryptionKey<MlKem1024>>h: Array<u8, UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>>