ml_kem

Module ml_kem_512

ML-KEM-512 is the parameter set for security category 1, corresponding to key search on a block cipher with a 128-bit key.

Structs

MlKem512

MlKem512 is the parameter set for security category 1, corresponding to key search on a block cipher with a 128-bit key.

Type Aliases

Ciphertext

Encoded ML-KEM-512 ciphertexts.

DecapsulationKey

An ML-KEM-512 DecapsulationKey which provides the ability to generate a new key pair, and decapsulate an encapsulated shared key.

EncapsulationKey

An ML-KEM-512 EncapsulationKey provides the ability to encapsulate a shared key so that it can only be decapsulated by the holder of the corresponding decapsulation key.