ml_kem::ml_kem_768

Type Alias DecapsulationKey

pub type DecapsulationKey = DecapsulationKey<MlKem768>;

An ML-KEM-768 DecapsulationKey which provides the ability to generate a new key pair, and decapsulate an encapsulated shared key.

Aliased Type

pub struct DecapsulationKey { /* private fields */ }