ml_kem::pkcs8

Trait EncodePrivateKey

pub trait EncodePrivateKey {
    // Required method
    fn to_pkcs8_der(&self) -> Result<SecretDocument, Error>;

    // Provided methods
    fn to_pkcs8_pem(
        &self,
        line_ending: LineEnding,
    ) -> Result<Zeroizing<String>, Error> { ... }
    fn write_pkcs8_der_file(&self, path: impl AsRef<Path>) -> Result<(), Error> { ... }
    fn write_pkcs8_pem_file(
        &self,
        path: impl AsRef<Path>,
        line_ending: LineEnding,
    ) -> Result<(), Error> { ... }
}

Serialize a private key object to a PKCS#8 encoded document.

Required Methods

fn to_pkcs8_der(&self) -> Result<SecretDocument, Error>

Serialize a SecretDocument containing a PKCS#8-encoded private key.

Errors

Returns format-specific errors in the event the document failed to serialize.

Provided Methods

fn to_pkcs8_pem( &self, line_ending: LineEnding, ) -> Result<Zeroizing<String>, Error>

Serialize this private key as PEM-encoded PKCS#8 with the given LineEnding.

Errors

Returns the same errors as EncodePrivateKey::to_pkcs8_der.

Returns the same errors as SecretDocument::to_pem.

fn write_pkcs8_der_file(&self, path: impl AsRef<Path>) -> Result<(), Error>

Write ASN.1 DER-encoded PKCS#8 private key to the given path.

Errors

Returns the same errors as EncodePrivateKey::to_pkcs8_der.

Returns errors in the event the file could not be written to the filesystem.

fn write_pkcs8_pem_file( &self, path: impl AsRef<Path>, line_ending: LineEnding, ) -> Result<(), Error>

Write ASN.1 PEM-encoded PKCS#8 private key to the given path.

Errors

Returns the same errors as EncodePrivateKey::to_pkcs8_der.

Returns errors in the event the file could not be written to the filesystem.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.