ml_kem::pkcs8

Trait EncodePublicKey

pub trait EncodePublicKey {
    // Required method
    fn to_public_key_der(&self) -> Result<Document, Error>;

    // Provided methods
    fn to_public_key_pem(
        &self,
        line_ending: LineEnding,
    ) -> Result<String, Error> { ... }
    fn write_public_key_der_file(
        &self,
        path: impl AsRef<Path>,
    ) -> Result<(), Error> { ... }
    fn write_public_key_pem_file(
        &self,
        path: impl AsRef<Path>,
        line_ending: LineEnding,
    ) -> Result<(), Error> { ... }
}

Serialize a public key object to a SPKI-encoded document.

Required Methods

fn to_public_key_der(&self) -> Result<Document, Error>

Serialize a Document containing a SPKI-encoded public key.

Errors

Returns encoding errors specific to the concrete type which impls this trait.

Provided Methods

fn to_public_key_pem(&self, line_ending: LineEnding) -> Result<String, Error>

Serialize this public key as PEM-encoded SPKI with the given LineEnding.

Errors

Returns encoding errors specific to the concrete type which impls this trait.

fn write_public_key_der_file(&self, path: impl AsRef<Path>) -> Result<(), Error>

Write ASN.1 DER-encoded public key to the given path.

Errors

Returns encoding errors specific to the concrete type which impls this trait.

fn write_public_key_pem_file( &self, path: impl AsRef<Path>, line_ending: LineEnding, ) -> Result<(), Error>

Write ASN.1 PEM-encoded public key to the given path.

Errors

Returns encoding errors specific to the concrete type which impls this trait.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.