ml_kem

Struct EncapsulationKey

pub struct EncapsulationKey<P>
where
    P: KemParams,
{
    ek_pke: MaybeBox<EncryptionKey<P>>,
    h: B32,
}

An EncapsulationKey provides the ability to encapsulate a shared key so that it can only be decapsulated by the holder of the corresponding decapsulation key.

Fields
ek_pke: MaybeBox<EncryptionKey<P>>h: B32
Implementations

impl<P> EncapsulationKey<P>
where P: Kem<SharedKeySize = U32> + KemParams,

pub fn new(encapsulation_key: &Key<Self>) -> Result<Self, InvalidKey>

Create a new EncapsulationKey from its serialized form.

Errors

If the key failed validation during decoding.

pub(crate) fn from_encryption_key(ek_pke: EncryptionKey<P>) -> Self

Convert from an EncryptionKey.

pub(crate) fn ek_pke(&self) -> &EncryptionKey<P>

Borrow the encryption key.

pub(crate) fn h(&self) -> B32

Retrieve the hash of the encryption key.