ml_kem

Trait KemCore

pub trait KemCore {
    type SharedKeySize: ArraySize;
    type CiphertextSize: ArraySize;
    type DecapsulationKey: Decapsulate<Ciphertext<Self>, SharedKey<Self>> + EncodedSizeUser + Debug + PartialEq;
    type EncapsulationKey: Encapsulate<Ciphertext<Self>, SharedKey<Self>> + EncodedSizeUser + Debug + PartialEq;

    // Required method
    fn generate(
        rng: &mut impl CryptoRngCore,
    ) -> (Self::DecapsulationKey, Self::EncapsulationKey);
}

A generic interface to a Key Encapsulation Method

Required Associated Types

type SharedKeySize: ArraySize

The size of a shared key generated by this KEM

type CiphertextSize: ArraySize

The size of a ciphertext encapsulating a shared key

type DecapsulationKey: Decapsulate<Ciphertext<Self>, SharedKey<Self>> + EncodedSizeUser + Debug + PartialEq

A decapsulation key for this KEM

type EncapsulationKey: Encapsulate<Ciphertext<Self>, SharedKey<Self>> + EncodedSizeUser + Debug + PartialEq

An encapsulation key for this KEM

Required Methods

fn generate( rng: &mut impl CryptoRngCore, ) -> (Self::DecapsulationKey, Self::EncapsulationKey)

Generate a new (decapsulation, encapsulation) key pair

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.