ml_kem

Trait TryKeyInit

pub trait TryKeyInit: Sized + KeySizeUser {
    // Required method
    fn new(key: &Array<u8, Self::KeySize>) -> Result<Self, InvalidKey>;

    // Provided method
    fn new_from_slice(key: &[u8]) -> Result<Self, InvalidKey> { ... }
}

Types which can be fallibly initialized from a key.

Required Methods

fn new(key: &Array<u8, Self::KeySize>) -> Result<Self, InvalidKey>

Create new value from a fixed-size key.

Errors

If the key is considered invalid according to rules specific to the implementing type.

Provided Methods

fn new_from_slice(key: &[u8]) -> Result<Self, InvalidKey>

Create new value from a variable size key.

Errors

If the key is considered invalid according to rules specific to the implementing type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.