ml_kem

Type Alias DecapsulationKey1024

pub type DecapsulationKey1024 = DecapsulationKey;

An ML-KEM-1024 DecapsulationKey which provides the ability to generate a new key pair, and decapsulate an encapsulated shared key.

Aliased Type

pub struct DecapsulationKey1024 {
    dk_pke: MaybeBox<DecryptionKey<MlKem1024>>,
    ek: EncapsulationKey<MlKem1024>,
    d: Option<MaybeBox<Array<u8, UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>>>>,
    z: MaybeBox<Array<u8, UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>>>,
}

Fields
dk_pke: MaybeBox<DecryptionKey<MlKem1024>>

Decryption key.

ek: EncapsulationKey<MlKem1024>

Associated encapsulation key.

d: Option<MaybeBox<Array<u8, UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>>>>

Seed this key was initialized from.

z: MaybeBox<Array<u8, UInt<UInt<UInt<UInt<UInt<UInt<UTerm, B1>, B0>, B0>, B0>, B0>, B0>>>

Random string used during the implicit rejection process.