ml_kem

Module algebra

Section 2.4. Interpreting the Pseudocode Section 4.2.2. Sampling algorithms Section 4.3. The Number-Theoretic Transform

Structs

FieldElement

An element of GF(q). Although q is only 16 bits wide, we use a wider uint type to so that we can defer modular reductions.

FieldElementReader 🔒

NttMatrix

A K x K matrix of NTT-domain polynomials. Each vector represents a row of the matrix, so that multiplying on the right just requires iteration.

NttPolynomial

An element of the ring T_q, i.e., a tuple of 128 elements of the direct sum components of T_q.

NttVector

A vector of K NTT-domain polynomials

Polynomial

An element of the ring R_q, i.e., a polynomial over Z_q of degree 255

PolynomialVector

A vector of polynomials of length k

Constants

GAMMA 🔒

ZETA_POW_BITREV 🔒

Type Aliases

Integer