ml_kem

Trait Encapsulate

pub trait Encapsulate: TryKeyInit + KeyExport {
    type Kem: Kem;

    // Required method
    fn encapsulate_with_rng<R>(
        &self,
        rng: &mut R,
    ) -> (Array<u8, <Self::Kem as Kem>::CiphertextSize>, Array<u8, <Self::Kem as Kem>::SharedKeySize>)
       where R: CryptoRng + ?Sized;
}

Encapsulator for shared secrets.

Often, this will just be a public key. However, it can also be a bundle of public keys, or it can include a sender’s private key for authenticated encapsulation.

Required Associated Types

type Kem: Kem

KEM algorithm this encapsulator is for.

Required Methods

fn encapsulate_with_rng<R>( &self, rng: &mut R, ) -> (Array<u8, <Self::Kem as Kem>::CiphertextSize>, Array<u8, <Self::Kem as Kem>::SharedKeySize>)
where R: CryptoRng + ?Sized,

Encapsulates a fresh SharedKey generated using the supplied random number generator R.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.